Add IMaybe interface
In this commit, the interface has no implementations, so it serves no
practical purpose. The motivation for the commit, though, is that you
can check out this commit and do the exercise of implementing the
interface.

There ought to be two implementations of IMaybe<T>: Just<T> and
Nothing<T>.