Add IEither<L, R> interface
This interface doesn't do anything in this commit, and has no
implementations. Thus, it can be used as an execercise for a reader
who'd like to try her or his hands on implementing the interfaces.

Another naming convention for this type is IResult<S, E>, where the S
stand for Success, and E for Error. This monad and bifunctor is,
however, more often known as Either, and the cases as Left and Right,
hence the generic type argument names of L and R. Notice that compared
to IResult<S, E>, the Right case is often used to indicate success
(because the correct behaviour is 'right'), leaving the Left case for
errors. Notice that this switches the role of the two dimensions, as in
IResult, S is to the left, and E to the right.